+++ cafeback/AroundCafe/src/main/java/com/example/demo/mypage/cafe/controller/CafeController.java
@@ -1,8 +1,7 @@
 package com.example.demo.mypage.cafe.controller;
 
-import com.example.demo.mypage.cafe.dto.CafeDto;
 import com.example.demo.mypage.cafe.entity.Cafe;
-import com.example.demo.mypage.cafe.service.CafeService;
+import com.example.demo.mypage.cafe.service.cafe.CafeService;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.MediaType;
@@ -42,12 +41,12 @@ public class CafeController {
                     writer.write(multipartFile.getBytes());
                     writer.close();
                 }
-                service.includeFileRegisterCafe(membNo,info,fileList);
+                service.includeFileModifyCafe(membNo,info,fileList);
             } catch (Exception e) {
                 return "register fail!";
             }
         }else if (fileList == null) {
-            service.notIncludeFileRegisterCafe(membNo,info);
+            service.notIncludeFileModifyCafe(membNo,info);
         }
 
         log.info("requestUpload(): Success!!!");
@@ -55,11 +54,8 @@ public class CafeController {
     }
 
     @GetMapping("/mypage/read/{membNo}")
-    public Cafe read(@PathVariable("membNo") Integer membNo) {
+    public Cafe myPageRead(@PathVariable("membNo") Integer membNo) {
         log.info("read Page no : "+ membNo);
-        return service.read(membNo);
-
-
+        return service.cafeMypageread(membNo);
     }
-
 }
