--- cafeback/AroundCafe/src/main/java/com/example/demo/mypage/cafe/repository/CafeImgRepository.java
@@ -1,7 +1,0 @@
-package com.example.demo.mypage.cafe.repository;
-
-import com.example.demo.mypage.cafe.entity.CafeImg;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface CafeImgRepository extends JpaRepository<CafeImg,Long> {
-}
