+++ cafeback/AroundCafe/src/main/resources/application.yaml
@@ -16,7 +16,7 @@ spring:
     show-sql: true
     generate-ddl: true
     hibernate:
-      ddl-auto: update
+      ddl-auto: create
 
   servlet:
     multipart:
